requirements.txt
# File: requirements.txt (in root directory)
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.2.3
Flask-SocketIO==5.1.1
psycopg2-binary==2.9.1
pika==1.2.0
pymongo==3.12.0
python-dotenv==0.19.0
psutil==5.8.0
requests==2.26.0
eventlet==0.30.2
python-engineio==4.2.1
python-socketio==5.4.0